hf-space-download/requirements.txt
fastapi
uvicorn
torch
diffusers
transformers
Pillow
safetensors
torchvision
accelerate
python-multipart
# xformers  # Disabled for Windows compatibility

